+++ store/routes.py
@@ -28,3 +28,11 @@ store_app.add_url_rule(
     view_func=store_view,
     methods=["GET", "PUT", "DELETE"],
 )
+
+store_app.add_url_rule(
+    "/stores/<store_id>/pets/",
+    view_func=store_view,
+    methods=[
+        "GET",
+    ],
+)
